Revert "Merge remote-tracking branch 'refs/remotes/origin/master' into boolean_fracture"
This reverts commit ecc10dbe29280b33a5cbfc95274ce0287d99b0e0.

+++ extern/carve/include/carve/pointset_impl.hpp
@@ -29,7 +29,7 @@ namespace carve {
   namespace point {
 
     inline size_t PointSet::vertexToIndex_fast(const Vertex *v) const {
-      return (size_t)(v - &vertices[0]);
+      return v - &vertices[0];
     }
 
   }
